pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "tb-progression-transcriptome-meta"
version = "0.1.0"
description = "Reproducible TB progression transcriptome meta-analysis pipeline"
readme = "README.md"
requires-python = ">=3.11"
authors = [{name = "tbmeta contributors"}]
dependencies = [
  "typer>=0.12",
  "PyYAML>=6.0",
  "pandas>=2.2",
  "numpy>=1.26",
  "scipy>=1.12",
  "scikit-learn>=1.4",
  "statsmodels>=0.14",
  "matplotlib>=3.8",
  "seaborn>=0.13",
  "requests>=2.32",
  "tenacity>=8.2",
  "joblib>=1.4",
  "pyarrow>=16",
  "streamlit>=1.40",
  "jinja2>=3.1",
  "gseapy>=1.1"
]

[project.optional-dependencies]
dev = [
  "pytest>=8.2",
  "pytest-cov>=5.0",
  "ruff>=0.6",
  "black>=24.8",
  "mypy>=1.11",
  "types-PyYAML",
  "types-requests"
]
openclaw = [
  "openclaw>=2026.2.28"
]
ml-optional = [
  "xgboost>=2.1",
  "shap>=0.46"
]
r-optional = [
  "rpy2>=3.5"
]
geo-optional = [
  "GEOparse>=2.0"
]

[project.scripts]
tbmeta = "tbmeta.cli:main"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]

[tool.black]
line-length = 130

[tool.ruff]
line-length = 130
src = ["src", "tests"]

[tool.ruff.lint]
select = ["E", "F", "I", "B", "UP", "N"]
ignore = ["E203"]

[tool.mypy]
python_version = "3.11"
warn_unused_configs = true
warn_redundant_casts = true
warn_unused_ignores = true
strict_optional = true
ignore_missing_imports = true
